Reduce warnings (#4381)
* Reduce warnings

* Use sbt-tpolecat

* Add missing compiler options

* Add doc scalac options

* Tune warnings

* split warn macro between 2.12 and 2.13

* Remove package-object-classes lint

* Update sbt-tpolecat version

* Remove skip-package

* Fix cast error

* Fix repl project scalac configuration

+++ scio-cassandra/cassandra3/src/it/scala/com/spotify/scio/cassandra/CassandraIT.scala
@@ -165,7 +165,7 @@ object CassandraIT {
       i.toShort,
       i,
       i.toLong,
-      JBigDecimal.valueOf(i),
+      JBigDecimal.valueOf(i.toLong),
       i.toFloat,
       i.toDouble,
       i % 2 == 0,
@@ -176,7 +176,7 @@ object CassandraIT {
       u1,
       u2,
       s"varchar$i",
-      BigInteger.valueOf(i),
+      BigInteger.valueOf(i.toLong),
       List(s"list$i").asJava,
       Set(s"set$i").asJava,
       Map(s"key$i" -> i).asJava
