+++ scio-elasticsearch/es8/src/main/scala/com/spotify/scio/elasticsearch/ElasticsearchIO.scala
@@ -45,7 +45,7 @@ final case class ElasticsearchIO[T](esOptions: ElasticsearchOptions) extends Sci
 
   /** Save this SCollection into Elasticsearch. */
   override protected def write(data: SCollection[T], params: WriteP): Tap[Nothing] = {
-    val shards = if (params.numOfShards >= 0) params.numOfShards else esOptions.nodes.size
+    val shards = if (params.numOfShards >= 0) params.numOfShards else esOptions.nodes.size.toLong
     val credentials = esOptions.usernameAndPassword.map { case (username, password) =>
       new UsernamePasswordCredentials(username, password)
     }
@@ -77,7 +77,7 @@ object ElasticsearchIO {
   object WriteParam {
     private[elasticsearch] val DefaultErrorFn: BulkExecutionException => Unit = m => throw m
     private[elasticsearch] val DefaultFlushInterval = Duration.standardSeconds(1)
-    private[elasticsearch] val DefaultNumShards = -1
+    private[elasticsearch] val DefaultNumShards = -1L
     private[elasticsearch] val DefaultMaxBulkRequestOperations = 3000
     private[elasticsearch] val DefaultMaxBulkRequestBytes = 5L * 1024L * 1024L
     private[elasticsearch] val DefaultMaxRetries = 3
